indentation, rename utils to finalizeAuth, rename authRoute, extract the session secret to the config
