+++ src/routes/auth/index.js
@@ -1,22 +1,24 @@
-import express  from 'express';
-import session  from 'express-session';
+import express               from 'express';
+import session               from 'express-session';
 import SequelizeStoreFactory from 'connect-session-sequelize';
 
-import basic    from './basic';
-import facebook from './facebook';
+import basic         from './basic';
+import facebook      from './facebook';
 
-import config   from '../../config';
-import { sequelize }       from '../../models/db';
+import config        from '../../config';
+import { sequelize } from '../../models/db';
 
 const router = express.Router();
 
+// We don't need the session handling for Basic authentication, that's why we
+// configure it here before inserting the session middleware.
 router.use('/basic', basic);
 
 // initalize sequelize with session store
 const SequelizeStore = SequelizeStoreFactory(session.Store);
 const isProduction = process.env.NODE_ENV === 'production';
 router.use(session({
-  secret: 'keyboard cat',
+  secret: config.get('userAuth.sessionSecret'),
   store: new SequelizeStore({
     db: sequelize
   }),
